pipecat-test/pyproject.toml
[project]
name = "pipecat-quickstart"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[deepgram,cartesia,ollama,silero,webrtc,runner]",
    "pipecat-ai-small-webrtc-prebuilt",
    # Local STT (Parakeet via onnx-asr)
    "onnx-asr[hub]>=0.9.0",
    # Local TTS (Chatterbox HTTP client)
    "aiohttp>=3.11.0",
]

[tool.uv.sources]
pipecat-ai-small-webrtc-prebuilt = { path = "../", editable = true }

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]